Added a hook to the round_start event to clear out all skeet info for all players and set all players to not be marked as currently pouncing (had some weirdness where after a few rounds, infected players were marked as having assisted in the skeet of another infected player... that's not right).
Updated the deadstop detection to clear out the skeet data for the hunter that got ds'd, as well as mark them not pouncing (they aren't going to be skeeted even if they get shot out of the air after the ds and haven't touched the ground yet).

Added some redundancy to the player death event to mark the hunter as not pouncing (can't pounce if you're dead, right?) just to make sure that the announcements don't accidentally have stale data.