Rearranged a lot of code to ensure that the tank events (spawned and killed) are hooked properly at:
1) The beginning of a round
2) The death of the tank
3) The end of the round

This keeps the tank hooks active instead of only showing damage done under certain (odd) circumstances.

Also, added in a variable to mark whether a tank had spawned at all this round and only prints tank damage (if enabled) when true; this value is only set to true in the tank spawn event.